+++ coop/src/main/java/Coop/coop/Services/RemarkService.java
@@ -2,6 +2,7 @@ package Coop.coop.Services;
 
 import Coop.coop.Entities.Remark;
 import Coop.coop.Interfaces.RemarkRepository;
+import Coop.coop.Interfaces.RemarkRepositoryCustom;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
@@ -11,18 +12,27 @@ import java.util.Optional;
 @Service
 public class RemarkService
 {
-    private final RemarkRepository remarkRepository;
+    private final RemarkRepositoryCustom remarkRepository;
 
     @Autowired
-    public RemarkService(RemarkRepository remarkRepository){
+    public RemarkService(RemarkRepositoryCustom remarkRepository){
         this.remarkRepository = remarkRepository;
     }
 
-    public Remark addRemark (Remark remark){
+    public Remark addRemark (Remark remark)
+    {
+        if(remark.getBody().isEmpty()){
+            throw new IllegalArgumentException("Your remark cannot be empty");
+        }
+
         return remarkRepository.save((remark));
     }
 
     public Boolean updateRemark(Remark remark){
+
+        if(remark.getBody().isEmpty()){
+            throw new IllegalArgumentException("Your remark cannot be empty");
+        }
 
         Remark oldRemark = new Remark();
         Optional<Remark> optionalRemark = remarkRepository.findById(remark.getId());
@@ -34,6 +44,7 @@ public class RemarkService
             oldRemark.setDateAdded(optionalRemark.get().getDateAdded());
             oldRemark.setStemNumber(optionalRemark.get().getStemNumber());
             oldRemark.setTimeInTrack(optionalRemark.get().getTimeInTrack());
+            oldRemark.setBody(optionalRemark.get().getBody());
 
             remarkRepository.save(oldRemark);
         }
@@ -43,11 +54,19 @@ public class RemarkService
         return true;
     }
 
-    public void deleteRemark(long remarkID){
-        remarkRepository.deleteById(remarkID);
+    public boolean deleteRemark(long remarkID){
+        if(remarkRepository.getById(remarkID) != null){
+            remarkRepository.deleteById(remarkID);
+            return true;
+        }
+        return false;
     }
     public List<Remark> getRemarksForSong(long songID){
 
-        return remarkRepository.findAll(Sort.by(Sort.Direction.ASC, "songID"));
+        return remarkRepository.findAllBySongID(songID);
+    }
+
+    public Optional<Remark> getRemark(long remarkID){
+        return remarkRepository.findById(remarkID);
     }
 }
